+++ src/likelihood.h
@@ -9,7 +9,7 @@ void Trajectory(double* times, double* traj_pars, double* d_arr,
 #ifdef __cplusplus
 extern "C" {
 #endif
-void Calculate_Lightcurve(double* times, long Nt, double* pars, double* template_);
+void Calculate_Lightcurve(double* times, size_t Nt, double* pars, double* template_);
 double Log_Likelihood(double all_sector_phases[], double all_sector_fluxes[],
                       double all_sector_uncertainties[], long int points_per_sector[],
                       const int NSECTORS, double all_parameters[],
